// src/models/set-type.ts
export enum SetType {
    // A yearly Magic core set (Tenth Edition, etc)
    core = "core",

    // A rotational expansion set in a block (Zendikar, etc)
    expansion = "expansion",

    // A reprint set that contains no new cards (Modern Masters, etc)
    masters = "masters",

    // Masterpiece Series premium foil cards
    masterpiece = "masterpiece",

    // From the Vault gift sets
    from_the_vault = "from_the_vault",

    // Premium Deck Series decks
    premium_deck = "premium_deck",

    // Duel Decks
    duel_deck = "duel_deck",

    // Commander preconstructed decks
    commander = "commander",

    // Planechase sets
    planechase = "planechase",

    // Conspiracy sets
    conspiracy = "conspiracy",

    // Archenemy sets
    archenemy = "archenemy",

    // Vanguard card sets
    vanguard = "vanguard",

    // A funny un-set or set with funny promos (Unglued, Happy Holidays, etc)
    funny = "funny",

    // A starter/introductory set (Portal, etc)
    starter = "starter",

    // A gift box set
    box = "box",

    // A set that contains purely promotional cards
    promo = "promo",

    // A set made up of tokens and emblems.
    token = "token",

    // A set made up of gold-bordered, oversize, or trophy cards that are not legal
    memorabilia = "memorabilia"
}